IA-AI/pyproject.toml
[project]
name = "ia-ai"
version = "0.1.0"
description = "智能农业 AI 服务"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.1.0",
    "langgraph-checkpoint-postgres>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "chromadb>=0.4.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "sentence-transformers>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []